Credit Page v3;
Created a new database to store the subject and assessment data in an SQL Table.
Used functions to execute different processes such as when the user enters their grade achieved and saving the input to the database. The assessment will also show when the subject is clicked via dropdown menu.