Credit Page v7;
The assessment data is now has its own script (allstandards.py) where when the script is run, all the standards avaliable are populated in the users.db database as a new table.
This is then connected to the credit page and check if the year level and subject code added together matches the standard/assessment subject code and then added to the GUI table.
Next steps is to fully have all standards shown and fit all columns on the screen.