Credit Page v9;
All columns of the table is shown and in the right place. Added functions to when the user enters their grade, it goes to a new table that is personalised to track their progress. Also started fetching that data to show in a pie chart.
Currently, when the user clicks save, the data gets entered in the table, but doesn't show on the GUI and the entry widget only goes away for the first standard only. The pie chart button also does not work when clicked. Only works when the user clicks save grades.