Credit Page v11;
Table overflow fixed, navigation bar, and the grades the users input gets locked and cannot be edited after saving. Though this only works when you enter the grades in the assessment order due to using a list.
Next step is adding a progress bar for NCEA Level certificate, endorsement, and table of credits with totals.