Course Search Page v5;
Added the bookmark feature through adding 3 new functions. First creating the frame where the bookmarked course is going to go, then adding the bookmark to a table in users.db and display, and last function is if the user wants to remove the bookmark, they click remove which would run the function to remove the course in the table and therefore removing it from the bookmark frame.