Landing Page v5;
Changing from Tkinter to CustomTkinter (CTk) because it looks more modern and up to date for my app to look more high quality and aestheically pleasing to look and use. Navigation button works through adding a for loop and running a function to destroy current window and run new window through subprocess.
Currently the widgets in the top and bottom frames do not display so fixing the frames is the next step.