Landing Page v6;
Trying to experiment with assigning the buttons to a frame for each individual page so it opens without having to destroy the current window and opening a new window. Each new page frame is a CTkFrame that is assign to main_frame which is the one that gets shown when the script is ran.