Landing Page v7;
Got the CTk widgets to work. All the Tk widgets from v4 are now converted to CTk and is now showing.
Next step is to wait for the other pages to finish in order to display the data from both creditpg and coursesearchpg since both are where the data is entered and stored.