Landing Page v8;
Modified the code so that it has the same navigation bar and GUI layout as the other 2 pages. Minor adjustments to the buttons to adapt.
Next steps is to split the page into 2 frames again. The top frame will have the progress bar, rank score table, and pie chart from fcreditpg.py. The bottom frame will have the bookmarked courses from fcoursesearchpg.py.