Landing Page v9;
Have now split the page into 4 frames.
Navigation frame is on the left.
Welcome frame is the welcome text which is at the very top.
Top frame is below welcome frame
Bottom frame is below top frame
Next step is to incorporate the bookmarked course function from fcoursesearchpg.py in the bottom frame.