ALL PAGES:
Made it so that all window will only open if there is the student_id in the database (which can only be entered when you register). All the data shown on the pages will be unique to them like what a personal account would be like.
I have done this through imprt sys and passing student_id variable through all the pages via the navigation buttons.